requirements.txt
faiss-cpu==1.8.0.post1
langchain-community==0.3.0
langchain-openai==0.2.1
langchain==0.3.1
marshmallow==3.22.0
openai==1.47.0
python-dotenv==1.0.1
sentence-transformers==3.1.1
